refactor rauthy-client and implement actix-web feature properly
