+++ rauthy-client/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "rauthy-client"
-version = "0.1.0-20231123"
+version = "0.1.0-20231127"
 edition = "2021"
 authors = ["Sebastian Dobe <sebastiandobe@mailbox.org>"]
 license = "Apache-2.0"
@@ -12,9 +12,19 @@ readme = "README.md"
 repository = "https://github.com/sebadob/rauthy/tree/main/rauthy-client"
 
 [features]
-default = ["axum"]
-actix-web = []
-axum = ["dep:axum", "dep:axum-extra", "dep:time"]
+default = []
+actix-web = [
+    "dep:actix-web",
+    "dep:http",
+]
+axum = [
+    "dep:axum",
+    "dep:axum-extra",
+
+    # minimal versions
+    "dep:bitflags",
+    "dep:elliptic-curve",
+]
 
 [dependencies]
 # common
@@ -35,12 +45,15 @@ serde_json = "1.0.100"
 tokio = "1.34"
 tracing = "0.1.40"
 
+# actix-web
+actix-web = { version = "4.4", optional = true, features = [] }
+http = { version = "1.0.0", optional = true }
+
 # axum
 axum = { version = "0.7", optional = true, features = [] }
 axum-extra = { version = "0.9", optional = true, features = ["cookie", "typed-header"] }
-time = { version = "0.3.30", optional = true, features = [] }
 
 # make minimal versions happy
 # both needed for axum 0.6.20
-bitflags = "1.3.2"
-elliptic-curve = "0.13.8"
+bitflags = { version = "1.3.2", optional = true }
+elliptic-curve = { version = "0.13.8", optional = true }
